backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg[binary]==3.2.13
pydantic-settings==2.10.1
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
python-jose==3.5.0
